+++ server/controllers/organism.js
@@ -3,6 +3,7 @@ const Organism = db.organism
 
 const AddOrganism = async(req, res)=>{
     const {body} = req
+    if(!body.name || !body.adress || body.phone) throw Error ("All fields are required")
     const organism = await Organism.create({
         ...body
     })
@@ -19,6 +20,8 @@ const getOneOrganism = async(req,res)=>{
 }
 const UpdateOrganism = async(req,res)=>{
     let id = req.params.id
+    const Org = await Organism.findById(id) 
+    if(!Org) throw Error('Organism not found')
     const body = req.body
     const UpdateOrganism = await Organism.findByIdAndUpdate(id,{
         ...body
@@ -29,6 +32,8 @@ const UpdateOrganism = async(req,res)=>{
 }
 const DeleteOrganism = async(req,res)=>{
     let id = req.params.id
+    const Org = await Organism.findById(id) 
+    if(!Org) throw Error('Organism not found')
     const deleteOrg = await Organism.findByIdAndDelete(id)
     if(!deleteOrg) throw Error('Error, try again')
     res.json({message: `Organism is deleted`, deleteOrg})
